requirements.txt
asgiref==3.6.0
click==8.1.3
Django==4.1.6
django-nested-admin==4.0.2
djangorestframework==3.14.0
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-monkey-business==1.0.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
Werkzeug==2.2.2
whitenoise==6.3.0
